# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laradisco API Documentation'
  description: "Internal reference for the Laradisco HTTP API. The API follows the JSON:API standard (https://jsonapi.org) and is built on Laravel's native JSON:API resources."
  version: 1.0.0
servers:
  - url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Obtain a token from <code>POST /api/v1/auth/login</code> (or <code>/auth/register</code>) and send it as <code>Authorization: Bearer &lt;token&gt;</code>.'
security:
  -
    default: []
paths:
...
  /api/v1/auth/reset-password:
    post:
      summary: 'Reset password'
      operationId: resetPassword
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\"message\": \"Your password has been reset successfully. You can now sign in.\"}\n\nReset the user's password using a valid reset code."
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address. Must not be greater than 255 characters.'
                  example: gbailey@example.net
                code:
                  type: string
                  description: 'Must be 6 characters.'
                  example: miyvdl
                password:
                  type: string
                  description: ''
                  example: '|]|{+-'
              required:
                - email
                - code
                - password
      security: []
...